c&c/component-2/requirements.txt
fastapi
strawberry-graphql[fastapi]
uvicorn
motor
python-dotenv